Add line breaks for function arguments
* Revise IntelliJ code style to wrap function arguments
if long, with a new line after ( and before ), and
aligning arguments at the continuation indent rather
than the end of the function.

This always moves arguments back to near the function
level, which helps when the arguments themselves
are long.

This example function:

```
void foo(int a, int b, int c) {
  //
}
```

was previously wrapped to:
```
void foo(int a,
         int b,
         int c) {

}
```

Now it becomes:
```
void foo(
    int a, int b,
    int c
) {
  //
}
```